ML: Conv2D + ConvTranspose2D accumulate in f32, not f64 (~2.4x faster SD, identical image)
Every Conv2DKernel variant (NCHW, NHWC, depthwise, native-low-p-weight) and ConvTranspose2DKernel accumulated
each output's inC*kH*kW MACs in double ("ultimate quality" caution). f64 is far slower on every GPU backend -
consumer cards run f64 at ~1/64 of f32, and WebGPU/WebGL EMULATE it via Dekker (multiple f32 ops per MAC) - so
the conv-heavy SD UNet/VAE paid it on every multiply-accumulate. f32 accumulation's rounding error over the MACs
is ~1e-5 relative, far below an 8-bit pixel and within the existing MAC-scaled conv-test tolerance.

Verified: Conv2D + ConvTranspose2D CPU-reference tests 164/0 on all 6 backends; SD-Turbo (seed 42) output is
perceptually identical to the f64 baseline (maxAbsDiff=1/255, meanAbsDiff=0.0007, 0.07% of bytes differ) and the
same generation went ~2.4x faster on CUDA (self-reported 29.4s -> 12.1s). The numerically-sensitive double
REDUCTIONS in Norm/Softmax/Stats (variance, softmax-sum) are intentionally left in f64; only conv MAC
accumulation changed. (Pipeline sweep: the only failures were transient model-download CONNECTION_RESETs - a
different Reference test each run, all failing in <150ms before any inference, not conv-value mismatches.)

+++ SpawnDev.ILGPU.ML/Conv2DKernel.cs
@@ -89,9 +89,12 @@ public class Conv2DKernel : IDisposable
         int oy = rem % outH;
         int oc = rem / outH;
 
-        // Double accumulation: eliminates precision errors across all backends.
-        // Always read bias — no branch (ANGLE optimizer workaround).
-        double sum = (double)bias[oc];
+        // f32 accumulation (the ML-standard for conv): the rounding error over the inC*kH*kW MACs is ~1e-5
+        // relative — imperceptible in an 8-bit image and within the MAC-scaled conv-test tolerance. f64
+        // accumulation here was over-cautious "ultimate quality" and is far slower on every GPU backend
+        // (consumer cards run f64 at ~1/64 of f32; WebGPU/WebGL EMULATE f64 via Dekker — the conv-heavy UNet/VAE
+        // paid that on every MAC). Always read bias — no branch (ANGLE optimizer workaround).
+        float sum = bias[oc];
 
         for (int ic = 0; ic < inC; ic++)
         {
@@ -107,18 +110,18 @@ public class Conv2DKernel : IDisposable
                     int ix = ox * stride + kx * dilationW - padLeft;
                     if (ix < 0 || ix >= inW) continue;
 
-                    sum += (double)input[icBase + iy * inW + ix] * (double)weight[wcBase + ky * kW + kx];
+                    sum += input[icBase + iy * inW + ix] * weight[wcBase + ky * kW + kx];
                 }
             }
         }
 
-        output[idx] = (float)sum;
+        output[idx] = sum;
     }
 
     /// <summary>
     /// Conv2D NCHW with NATIVE low-precision WEIGHTS (<typeparamref name="T"/> = ILGPU.Half / BFloat16 /
     /// Float8E*) — identical math to Conv2DImpl, but each filter weight is read NATIVELY and converted to
-    /// float in-register (PrecisionConvert) for the double-precision accumulation. The weight stays native in
+    /// float in-register (PrecisionConvert) for the f32 accumulation. The weight stays native in
     /// GPU memory (no f32 temp buffer); input/bias/output stay fp32, no accuracy loss. The UNet is mostly
     /// Conv, so this is the bulk of the low-p memory win for SD-Turbo.
     /// </summary>
@@ -140,7 +143,7 @@ public class Conv2DKernel : IDisposable
         int oy = rem % outH;
         int oc = rem / outH;
 
-        double sum = (double)bias[oc];
+        float sum = bias[oc];
 
         for (int ic = 0; ic < inC; ic++)
         {
@@ -156,12 +159,12 @@ public class Conv2DKernel : IDisposable
                     int ix = ox * stride + kx * dilationW - padLeft;
                     if (ix < 0 || ix >= inW) continue;
 
-                    sum += (double)input[icBase + iy * inW + ix] * (double)PrecisionConvert.ConvertToSingle(weight[wcBase + ky * kW + kx]);
+                    sum += input[icBase + iy * inW + ix] * PrecisionConvert.ConvertToSingle(weight[wcBase + ky * kW + kx]);
                 }
             }
         }
 
-        output[idx] = (float)sum;
+        output[idx] = sum;
     }
 
     /// <summary>
@@ -326,7 +329,7 @@ public class Conv2DKernel : IDisposable
         int oy = rem % outH;
         int c = rem / outH;
 
-        double sum = (double)bias[c];
+        float sum = bias[c];
 
         int inBase = c * inH * inW;
         int wBase = c * kH * kW;
@@ -340,11 +343,11 @@ public class Conv2DKernel : IDisposable
                 int ix = ox * stride + kx * dilationW - padLeft;
                 if (ix < 0 || ix >= inW) continue;
 
-                sum += (double)input[inBase + iy * inW + ix] * (double)weight[wBase + ky * kW + kx];
+                sum += input[inBase + iy * inW + ix] * weight[wBase + ky * kW + kx];
             }
         }
 
-        output[idx] = (float)sum;
+        output[idx] = sum;
     }
 
     public void ForwardDepthwise(
@@ -411,7 +414,7 @@ public class Conv2DKernel : IDisposable
         int ox = rem % outW;
         int oy = rem / outW;
 
-        double sum = (double)bias[oc];
+        float sum = bias[oc];
 
         int kernelSize = inC * kH * kW;
         for (int k = 0; k < kernelSize; k++)
@@ -428,10 +431,10 @@ public class Conv2DKernel : IDisposable
 
             int inIdx = (iy * inW + ix) * inC + ic;
             int wIdx = ((oc * kH + ky) * kW + kx) * inC + ic;
-            sum += (double)input[inIdx] * (double)weight[wIdx];
+            sum += input[inIdx] * weight[wIdx];
         }
 
-        output[idx] = (float)sum;
+        output[idx] = sum;
     }
 
     public void ForwardNHWC(
@@ -491,7 +494,7 @@ public class Conv2DKernel : IDisposable
         int ox = rem % outW;
         int oy = rem / outW;
 
-        double sum = (double)bias[c];
+        float sum = bias[c];
 
         int kernelSize = kH * kW;
         for (int k = 0; k < kernelSize; k++)
@@ -505,10 +508,10 @@ public class Conv2DKernel : IDisposable
 
             int inIdx = (iy * inW + ix) * C + c;
             int wIdx = (ky * kW + kx) * C + c;
-            sum += (double)input[inIdx] * (double)weight[wIdx];
+            sum += input[inIdx] * weight[wIdx];
         }
 
-        output[idx] = (float)sum;
+        output[idx] = sum;
     }
 
     public void ForwardDepthwiseNHWC(
